split in 2 ugly files. Shift q: quit
